# backend/openapi/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Knowledge Base External API",
        "version": "0.1.0",
        "description": "External HTTP API (G9). 支持 JWT 或 API Key 鉴权，并复用部门范围过滤逻辑。"
    },
    "servers": [
        {
            "url": "http://localhost:3001"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key"
            }
        },
        "schemas": {
            "UnifiedResponse": {
                "type": "object",
                "required": ["code", "msg", "data"],
                "properties": {
                    "code": { "type": "integer" },
                    "msg": { "type": "string" },
                    "data": { "type": "object" }
                }
            },
            "SearchRequest": {
                "type": "object",
                "required": ["query"],
                "properties": {
                    "query": { "type": "string" },
                    "topK": { "type": "integer", "minimum": 1, "maximum": 50, "default": 8 },
                    "hybrid": { "type": "boolean", "default": true, "description": "是否混合关键字分支（scroll + RRF）；false 为纯向量" },
                    "includeAnswer": { "type": "boolean", "default": true, "description": "是否调用 vLLM 生成答复" },
                    "includeRag": { "type": "boolean", "description": "兼容旧字段，语义同 includeAnswer" },
                    "includeHits": { "type": "boolean", "default": true, "description": "是否在 data 中返回检索片段 hits；false 时仍内部检索并供 RAG 使用，响应更精简" }
                }
            },
...
        }
    },
    "security": [
        { "bearerAuth": [] },
        { "apiKeyAuth": [] }
    ],
    "paths": {
...
        "/api/v1/documents": {
            "get": {
                "tags": ["Documents"],
                "summary": "文档列表（分页，不含已删除）",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": { "type": "integer", "minimum": 1, "default": 1 }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/UnifiedResponse" }
                            }
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/UnifiedResponse" }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/search": {
            "post": {
                "tags": ["Search"],
                "summary": "混合检索（向量 + 关键字 scroll/RRF，可 hybrid=false 关闭）。默认启用 vLLM 生成 Markdown 答复（includeAnswer=false 或 includeRag=false 可关闭）。",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SearchRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        { "$ref": "#/components/schemas/UnifiedResponse" }
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/UnifiedResponse" }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (JWT/API Key missing or invalid)",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/UnifiedResponse" }
                            }
                        }
                    },
                    "429": {
                        "description": "检索请求过频（IP 或 API Key 配额）",
                        "headers": {
                            "Retry-After": {
                                "schema": { "type": "integer" },
                                "description": "建议等待秒数"
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/UnifiedResponse" }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/UnifiedResponse" }
                            }
                        }
                    }
                }
            }
        },
...
    }
}